@@ -9,8 +9,8 @@
     <version>1.0-SNAPSHOT</version>
 
     <properties>
-        <maven.compiler.source>11</maven.compiler.source>
-        <maven.compiler.target>11</maven.compiler.target>
+        <maven.compiler.source>15</maven.compiler.source>
+        <maven.compiler.target>15</maven.compiler.target>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <appium.version>8.3.0</appium.version>
         <selenium.version>4.7.1</selenium.version>
@@ -20,6 +20,7 @@
         <logback.contrib.version>0.1.5</logback.contrib.version>
         <jackson.version>2.14.1</jackson.version>
         <spring.version>5.3.23</spring.version>
+        <owner.version>1.0.12</owner.version>
         <downloadSources>true</downloadSources>
         <downloadJavadocs>true</downloadJavadocs>
     </properties>
@@ -33,6 +34,12 @@
 
 
     <dependencies>
+
+        <dependency>
+            <groupId>org.aeonbits.owner</groupId>
+            <artifactId>owner</artifactId>
+            <version>${owner.version}</version>
+        </dependency>
 
         <dependency>
             <groupId>org.springframework</groupId>
@@ -381,6 +388,12 @@
             <version>2.1.5</version>
         </dependency>
 
+        <dependency>
+            <groupId>org.ahocorasick</groupId>
+            <artifactId>ahocorasick</artifactId>
+            <version>0.6.3</version>
+        </dependency>
+
     </dependencies>
 
     <build>
@@ -392,7 +405,9 @@
                 <artifactId>maven-compiler-plugin</artifactId>
                 <version>3.8.1</version>
                 <configuration>
-                    <release>11</release>
+                    <release>15</release>
+                    <source>15</source>
+                    <target>15</target>
                 </configuration>
             </plugin>
 
