@@ -3,6 +3,9 @@ package org.automation.elements;
 import io.appium.java_client.pagefactory.AppiumFieldDecorator;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.support.PageFactory;
+
+import java.lang.reflect.InvocationTargetException;
+
 public class ObjectFactoryGenerator extends PageFactory {
     public static <T extends ObjectFactoryGenerator> T instantiateWebPage(WebDriver driver, Class<T> pageClass) {
         try {
@@ -24,9 +27,18 @@ public class ObjectFactoryGenerator extends PageFactory {
     @Deprecated
     public static <T> T instantiateObject(Class<T> pageClass) {
         try {
-            return pageClass.newInstance();
+            return pageClass.getConstructor().newInstance();
         } catch (Exception exception) {
             throw new RuntimeException(exception);
         }
     }
+    @Deprecated
+    public static <T> T instantiateObject(Class<T> pageClass, Class<?> parametersTypes) {
+        try {
+            return pageClass.getConstructor(parametersTypes).newInstance(new Object());
+        } catch (Exception exception) {
+            throw new RuntimeException(exception);
+        }
+    }
+
 }
