@@ -1,8 +1,7 @@
 package org.extensions;
 
 import org.extensions.anontations.JacksonProvider;
-import org.files.jsonReader.JacksonHelperExtension;
-import org.junit.jupiter.api.Assertions;
+import org.files.jsonReader.JacksonExtension;
 import org.junit.jupiter.api.extension.ExtensionContext;
 import org.junit.jupiter.api.extension.ParameterContext;
 import org.junit.jupiter.api.extension.ParameterResolver;
@@ -10,7 +9,7 @@ import java.util.Optional;
 
 public class JacksonProviderExtension implements ParameterResolver {
 
-    private final ThreadLocal<JacksonHelperExtension<?>> jacksonHelper = new ThreadLocal<>();
+    private final ThreadLocal<JacksonExtension<?>> jacksonHelper = new ThreadLocal<>();
 
     @Override
     public synchronized boolean supportsParameter(ParameterContext parameter, ExtensionContext context) {
@@ -23,7 +22,7 @@ public class JacksonProviderExtension implements ParameterResolver {
             Optional<JacksonProvider> provider = Optional.ofNullable(context.getElement().get().getAnnotation(JacksonProvider.class));
             if (provider.isPresent()) {
                 final String path = System.getProperty("user.dir") + "/" + provider.get().dir();
-                this.jacksonHelper.set(new JacksonHelperExtension<>(path, provider.get().fileName() , provider.get().classObject()));
+                this.jacksonHelper.set(new JacksonExtension<>(path, provider.get().fileName() , provider.get().classObject()));
                 return this.jacksonHelper.get();
             } else throw new RuntimeException("fail resolve jackson provider initiation");
         }
