@@ -30,7 +30,7 @@ public class MobileBasePocTest {
 
     @Test
     @Repeat(onStatus = { FAIL, SKIP })
-    @DriverJsonProvider(proxyPort = 0, jsonCapsPath = "android.caps.json.first")
+    @DriverJsonProvider(proxyPort = 0, jsonCapsPath = "android.caps.json")
     @TestReportInfo(testId = 1, assignCategory = "poc", assignAuthor = "aviad", info = "MobileBasePocTest")
     public void a_mobileTest(@Autowired MobileDriverManager driverManager) {
         driverManager.click(elementToBeClickable(By.id("com.ideomobile.hapoalim:id/login_user_name_view_automation")));
@@ -39,7 +39,7 @@ public class MobileBasePocTest {
 
     @Test
     @Repeat(onStatus = { FAIL, SKIP })
-    @DriverJsonProvider(proxyPort = 0, jsonCapsPath = "android.caps.json.second")
+    @DriverJsonProvider(proxyPort = 0, jsonCapsPath = "android.caps.json")
     @TestReportInfo(testId = 1, assignCategory = "poc", assignAuthor = "aviad", info = "MobileBasePocTest")
     public void b_mobileTest(@Autowired MobileDriverManager driverManager) {
         driverManager.click(elementToBeClickable(By.id("com.ideomobile.hapoalim:id/login_user_name_view_automation")));
